requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
apache-airflow==2.8.1
black==24.3.0
ruff==0.3.4
bandit==1.7.8
pytest==8.1.1
